+++ native.js
@@ -21,7 +21,7 @@ const bindingsByPlatformAndArch = {
 		arm: { base: 'linux-arm-gnueabihf', musl: 'linux-arm-musleabihf' },
 		arm64: { base: 'linux-arm64-gnu', musl: 'linux-arm64-musl' },
 		loong64: { base: 'linux-loongarch64-gnu', musl: null },
-		ppc64: { base: 'linux-powerpc64le-gnu', musl: null },
+		ppc64: { base: 'linux-ppc64-gnu', musl: null },
 		riscv64: { base: 'linux-riscv64-gnu', musl: 'linux-riscv64-musl' },
 		s390x: { base: 'linux-s390x-gnu', musl: null },
 		x64: { base: 'linux-x64-gnu', musl: 'linux-x64-musl' }