+++ rust/bindings_napi/src/lib.rs
@@ -1,15 +1,15 @@
-use napi::bindgen_prelude::*;
+use napi::{bindgen_prelude::*, ScopedTask};
 use napi_derive::napi;
 use parse_ast::parse_ast;
 
 #[cfg(all(
-  not(all(target_os = "linux", target_env = "musl", target_arch = "aarch64")),
   not(all(target_os = "linux", target_arch = "loongarch64")),
+  not(all(target_os = "linux", target_arch = "riscv64", target_env = "musl")),
   not(all(target_os = "freebsd", target_arch = "aarch64")),
   not(debug_assertions)
 ))]
 #[global_allocator]
-static ALLOC: mimalloc_rust::GlobalMiMalloc = mimalloc_rust::GlobalMiMalloc;
+static ALLOC: mimalloc_safe::MiMalloc = mimalloc_safe::MiMalloc;
 
 pub struct ParseTask {
   pub code: String,
@@ -18,9 +18,9 @@ pub struct ParseTask {
 }
 
 #[napi]
-impl Task for ParseTask {
-  type Output = Buffer;
-  type JsValue = Buffer;
+impl<'task> ScopedTask<'task> for ParseTask {
+  type Output = Vec<u8>;
+  type JsValue = BufferSlice<'task>;
 
   fn compute(&mut self) -> Result<Self::Output> {
     Ok(
@@ -33,14 +33,19 @@ impl Task for ParseTask {
     )
   }
 
-  fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
-    Ok(output)
+  fn resolve(&mut self, env: &'task Env, output: Self::Output) -> Result<Self::JsValue> {
+    BufferSlice::from_data(env, output)
   }
 }
 
 #[napi]
-pub fn parse(code: String, allow_return_outside_function: bool, jsx: bool) -> Buffer {
-  parse_ast(code, allow_return_outside_function, jsx).into()
+pub fn parse(
+  env: &Env,
+  code: String,
+  allow_return_outside_function: bool,
+  jsx: bool,
+) -> Result<BufferSlice> {
+  BufferSlice::from_data(env, parse_ast(code, allow_return_outside_function, jsx))
 }
 
 #[napi]
@@ -61,16 +66,16 @@ pub fn parse_async(
 }
 
 #[napi]
-pub fn xxhash_base64_url(input: Uint8Array) -> String {
-  xxhash::xxhash_base64_url(&input)
+pub fn xxhash_base64_url(input: &[u8]) -> String {
+  xxhash::xxhash_base64_url(input)
 }
 
 #[napi]
-pub fn xxhash_base36(input: Uint8Array) -> String {
-  xxhash::xxhash_base36(&input)
+pub fn xxhash_base36(input: &[u8]) -> String {
+  xxhash::xxhash_base36(input)
 }
 
 #[napi]
-pub fn xxhash_base16(input: Uint8Array) -> String {
-  xxhash::xxhash_base16(&input)
+pub fn xxhash_base16(input: &[u8]) -> String {
+  xxhash::xxhash_base16(input)
 }