+++ src/ast/nodes/BinaryExpression.ts
@@ -12,11 +12,23 @@ import {
 	SHARED_RECURSION_TRACKER,
 	UNKNOWN_PATH
 } from '../utils/PathTracker';
+import { getRenderedLiteralValue } from '../utils/renderLiteralValue';
+import ExternalVariable from '../variables/ExternalVariable';
+import NamespaceVariable from '../variables/NamespaceVariable';
 import ExpressionStatement from './ExpressionStatement';
 import type { LiteralValue } from './Literal';
 import type * as NodeType from './NodeType';
-import { type LiteralValueOrUnknown, UnknownValue } from './shared/Expression';
-import { doNotDeoptimize, type ExpressionNode, NodeBase } from './shared/Node';
+import {
+	type InclusionOptions,
+	type LiteralValueOrUnknown,
+	UnknownValue
+} from './shared/Expression';
+import {
+	doNotDeoptimize,
+	type ExpressionNode,
+	type IncludeChildren,
+	NodeBase
+} from './shared/Node';
 
 type Operator =
 	| '!='
@@ -71,13 +83,18 @@ const binaryOperators: Partial<
 	// instanceof: () => UnknownValue,
 };
 
+const UNASSIGNED = Symbol('Unassigned');
+
 export default class BinaryExpression extends NodeBase implements DeoptimizableEntity {
 	declare left: ExpressionNode;
 	declare operator: keyof typeof binaryOperators;
 	declare right: ExpressionNode;
 	declare type: NodeType.tBinaryExpression;
+	renderedLiteralValue: string | typeof UnknownValue | typeof UNASSIGNED = UNASSIGNED;
 
-	deoptimizeCache(): void {}
+	deoptimizeCache(): void {
+		this.renderedLiteralValue = UnknownValue;
+	}
 
 	getLiteralValueAtPath(
 		path: ObjectPath,
@@ -88,6 +105,13 @@ export default class BinaryExpression extends NodeBase implements DeoptimizableE
 		const leftValue = this.left.getLiteralValueAtPath(EMPTY_PATH, recursionTracker, origin);
 		if (typeof leftValue === 'symbol') return UnknownValue;
 
+		// Optimize `'export' in namespace`
+		if (this.operator === 'in' && this.right.variable instanceof NamespaceVariable) {
+			const [variable] = this.right.variable.context.traceExport(String(leftValue));
+			if (variable instanceof ExternalVariable) return UnknownValue;
+			return !!variable;
+		}
+
 		const rightValue = this.right.getLiteralValueAtPath(EMPTY_PATH, recursionTracker, origin);
 		if (typeof rightValue === 'symbol') return UnknownValue;
 
@@ -95,6 +119,18 @@ export default class BinaryExpression extends NodeBase implements DeoptimizableE
 		if (!operatorFunction) return UnknownValue;
 
 		return operatorFunction(leftValue, rightValue);
+	}
+
+	getRenderedLiteralValue() {
+		// Only optimize `'export' in ns`
+		if (this.operator !== 'in' || !(this.right.variable instanceof NamespaceVariable)) {
+			return UnknownValue;
+		}
+
+		if (this.renderedLiteralValue !== UNASSIGNED) return this.renderedLiteralValue;
+		return (this.renderedLiteralValue = getRenderedLiteralValue(
+			this.getLiteralValueAtPath(EMPTY_PATH, SHARED_RECURSION_TRACKER, this)
+		));
 	}
 
 	hasEffects(context: HasEffectsContext): boolean {
@@ -113,9 +149,21 @@ export default class BinaryExpression extends NodeBase implements DeoptimizableE
 		return type !== INTERACTION_ACCESSED || path.length > 1;
 	}
 
+	include(
+		context: InclusionContext,
+		includeChildrenRecursively: IncludeChildren,
+		options?: InclusionOptions
+	) {
+		if (!this.included) this.includeNode(context);
+		if (typeof this.getRenderedLiteralValue() === 'symbol') {
+			this.left.include(context, includeChildrenRecursively, options);
+			this.right.include(context, includeChildrenRecursively, options);
+		}
+	}
+
 	includeNode(context: InclusionContext) {
 		this.included = true;
-		if (this.operator === 'in') {
+		if (this.operator === 'in' && typeof this.getRenderedLiteralValue() === 'symbol') {
 			this.right.includePath(UNKNOWN_PATH, context);
 		}
 	}
@@ -129,8 +177,13 @@ export default class BinaryExpression extends NodeBase implements DeoptimizableE
 		options: RenderOptions,
 		{ renderedSurroundingElement }: NodeRenderOptions = BLANK
 	): void {
-		this.left.render(code, options, { renderedSurroundingElement });
-		this.right.render(code, options);
+		const renderedLiteralValue = this.getRenderedLiteralValue();
+		if (typeof renderedLiteralValue !== 'symbol') {
+			code.overwrite(this.start, this.end, renderedLiteralValue);
+		} else {
+			this.left.render(code, options, { renderedSurroundingElement });
+			this.right.render(code, options);
+		}
 	}
 }