+++ test/form/samples/namespace-optimization-in-operator/main.js
@@ -1,0 +1,5 @@
+import * as foo from './foo';
+
+if ('d' in foo) console.log(foo.d());
+if ('c' in foo) console.log(foo.c());
+console.log('c' in foo);