+++ test/function/samples/exports-are-not-defined/main.js
@@ -1,0 +1,2 @@
+import { foo } from './module';
+console.log(foo);